+++ docker_daemon/CHANGELOG.md
@@ -1,10 +1,16 @@
 # CHANGELOG - docker_daemon
 
-1.3.1 / Unreleased
+1.3.2 / Unreleased
 ==================
 ### Changes
 
 * [BUGFIX] safely check volume list before accessing. See [#544][]
+
+1.3.1 / 2017-07-26 
+==================
+### Changes
+
+* [BUGFIX] fix event collection on ecs and nomad. See [#616][]
 
 1.3.0 / 2017-07-18
 ==================
@@ -75,4 +81,5 @@
 [#509]: https://github.com/DataDog/integrations-core/issues/509
 [#544]: https://github.com/DataDog/integrations-core/issues/544
 [#553]: https://github.com/DataDog/integrations-core/issues/553
+[#616]: https://github.com/DataDog/integrations-core/issues/616
 [@sophaskins]: https://github.com/sophaskins