+++ varnish/CHANGELOG.md
@@ -1,11 +1,19 @@
 # CHANGELOG - varnish
 
+1.0.4 Unreleased
+==================
+
+### Changes
+
+* [IMPROVEMENT] Support for passing additional parameters to varnishstat
+and varnishadm in order to better support service discovery.
+
 1.0.3 / 2017-07-18
 ==================
 
 ### Changes
 
-* [BUGFIX] Fixes an issue with retrieving the backend service checks. [#582][]
+* [BUGFIX] Fixes an issue with retrieving the backend service checks. Special thanks to @adongy for finding this! [#582][]
 
 1.0.2 / 2017-07-18
 ==================